Implemented dash lines (WIP), fix padding for Paragraph.
Implemented a working dash lines that goes through every circle. It fine but it's very hard coded and not really flexible (changing screen size will mess up the proportion). Also with the dash lines, not every steps (container for each wrapper) have the same amount of lines and correct spacing. Although it does work and is visually ok, it's not perfect. I then added padding to indent the paragraph to compensate for the lines going through the middle of the circle.