DDD-primary (added acents, default colors, etc.)
Implemented ddd-primary with added accents to the necessary CSS. Can change ddd-primary and it will respond with a color change. Set default color to ddd-primary=7 which is Nittany navy. MAYBE WIP (Not sure yet with the requirement, but might make ddd-primary reactive, EX: changing ddd-primary in inspect elemeents will change it live, Instead of needing to change it from the backend).